Move test code into testing module
