+++ temporal-sdk/src/main/java/io/temporal/failure/FailureConverter.java
@@ -37,7 +37,6 @@ import io.temporal.client.ActivityCanceledException;
 import io.temporal.common.converter.DataConverter;
 import io.temporal.common.converter.EncodedValues;
 import io.temporal.internal.common.CheckedExceptionWrapper;
-import io.temporal.testing.SimulatedTimeoutFailure;
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.util.Optional;
