+++ temporal-sdk/src/main/java/io/temporal/internal/sync/DeterministicRunnerWrapper.java
@@ -19,17 +19,21 @@
 
 package io.temporal.internal.sync;
 
+import com.google.common.annotations.VisibleForTesting;
+
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
 import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 
-class DeterministicRunnerWrapper implements InvocationHandler {
+@VisibleForTesting
+public class DeterministicRunnerWrapper implements InvocationHandler {
 
   private final InvocationHandler invocationHandler;
 
-  DeterministicRunnerWrapper(InvocationHandler invocationHandler) {
+  @VisibleForTesting
+  public DeterministicRunnerWrapper(InvocationHandler invocationHandler) {
     this.invocationHandler = Objects.requireNonNull(invocationHandler);
   }
