+++ temporal-sdk/src/test/java/io/temporal/workflow/LocalAsyncCompletionWorkflowTest.java
@@ -27,6 +27,7 @@ import io.temporal.activity.ActivityOptions;
 import io.temporal.client.ActivityCompletionClient;
 import io.temporal.client.WorkflowOptions;
 import io.temporal.common.RetryOptions;
+import io.temporal.testing.TestWorkflowRule;
 import io.temporal.worker.WorkerOptions;
 import java.time.Duration;
 import java.util.ArrayList;
